reworking the bindings and adding instructions
