--- sdk/go/postgres/db_table.go
@@ -1,12 +1,0 @@
-// generated by wit-bindgen 0.3.0
-package postgres
-
-type Column struct {
-	Name     string
-	DataType DbDataType
-}
-
-type RowSet struct {
-	Columns []Column
-	Rows    [][]DbValue
-}
