+++ sdk/go/postgres/postgres.go
@@ -1,20 +1,17 @@
 package postgres
 
 // #include "outbound-pg.h"
-// #include<stdlib.h>
+// #include <stdlib.h>
 import "C"
 
 import "unsafe"
 
 import "fmt"
 
-import "errors"
-
 func lowerPgStr(x string) C.outbound_pg_string_t {
 	return C.outbound_pg_string_t{ptr: C.CString(x), len: C.size_t(len(x))}
 }
 
-// generated by wit-bindgen 0.3.0
 func lowerParameterValues(params []ParameterValue) C.outbound_pg_list_parameter_value_t {
 	var lower_params C.outbound_pg_list_parameter_value_t
 	if len(params) == 0 {
@@ -148,7 +145,6 @@ func lowerParameterValues(params []ParameterValue) C.outbound_pg_list_parameter_
 	return lower_params
 }
 
-// generated by wit-bindgen 0.3.0
 func liftPgError(err *C.outbound_pg_pg_error_t) error {
 	var gstr string
 	switch int(err.tag) {
@@ -164,10 +160,9 @@ func liftPgError(err *C.outbound_pg_pg_error_t) error {
 	default:
 		gstr = fmt.Sprintf("unrecognized error: %v", err.tag)
 	}
-	return errors.New(gstr)
+	return fmt.Errorf(gstr)
 }
 
-// generated by wit-bindgen 0.3.0
 func liftRowSet(rowset *C.outbound_pg_row_set_t) RowSet {
 	var lift_rowset RowSet
 	lift_rowset_Columns := make([]Column, rowset.columns.len)
