+++ sdk/go/readme.md
@@ -19,3 +19,102 @@ func init() {
 
 func main() {}
 ```
+
+## Using wit-bindgen v0.4.0 to generate Go types
+
+As a shortcut to implementation, it is possible to use wit-bindgen v0.4.0 to generate the Go structs and helper functions however you will still need to use the wit-bindgen v0.2.0 to generate the C code. This is not a completely automoated solution but it does speed things up by not needing to write structs, enums and basic marshalling code.
+
+### Installing wit-bindgen v0.4.0
+
+To use the tinygo generator for wit-bindgen, you need to have wit-bindgen v4.0 on your path. In order to not conflict with existing installations, you can use this script to download locally.
+
+```shell
+VERSION="0.4.0"
+OS="macos"
+ARCH="aarch64"
+NAME="wit-bindgen-v${VERSION}-${ARCH}-${OS}"
+
+# download the release binary
+wget "https://github.com/bytecodealliance/wit-bindgen/releases/download/wit-bindgen-cli-${VERSION}/${NAME}.tar.gz" -O - | tar -xv --strip-components 1 "${NAME}/wit-bindgen"
+
+# move it to somewhere on your path with a non-conflicting name
+mv wit-bindgen ~/.local/bin/wit-bindgen4
+```
+
+### Modifying WIT for worlds
+
+Next you need to modify existing WIT files to be compatible with newer world syntax. Essentially, for each dependency you can create a new file and wrap it in `default interface x` and for the root dependency wrap it in `default world x`. For example, this is how you could modify WIT for `wit/ephemeral/outbound-pg.wit`.
+
+```shell
+mkdir wit/ephemeral/outbound-pg
+
+# use statement in wit/ephemeral/outbound-pg.wit
+cat << EOF > wit/ephemeral/outbound-pg/world.wit
+interface pg-types {
+  $(cat wit/ephemeral/pg-types.wit)
+}
+
+// NOTE: use statement in wit/ephemeral/outbound-pg.wit
+interface rdbms-types {
+  $(cat wit/ephemeral/rdbms-types.wit)
+}
+
+// NOTE: recreate wit/ephemeral/outbound-pg.wit
+default world outbound-pg {
+  // NOTES:
+  //  - the name of the world becomes the go package name
+  //  - dependecies from 'use' will have the interface name as a prefix
+  //  - use statement require specific types (no wildcards)
+  //  - add 'import' before function
+  //  - expected -> result
+
+  use self.pg-types.{pg-error}
+  use self.rdbms-types.{parameter-value, row-set}
+
+  // query the database: select
+  import query: func(address: string, statement: string, params: list<parameter-value>) -> result<row-set, pg-error>
+
+  // execute command to the database: insert, update, delete
+  import execute: func(address: string, statement: string, params: list<parameter-value>) -> result<u64, pg-error>
+}
+EOF
+```
+
+### Using the `tiny-go` generator
+
+To run the generator using local file dependencies, you must use the folder as the path. It will use the default world found in the WIT files. For example, this is how you could generate bindings for postgres:
+
+```
+# use the tiny-go generator
+wit-bindgen4 tiny-go --out-dir ./wit/ephemeral/outbound-pg/target/tiny-go/ ./wit/ephemeral/outbound-pg/world.wit
+
+# move only the basic outbound-pg.go file
+cp ./wit/ephemeral/outbound-pg/target/tiny-go/outbound-pg.go ./sdk/go/postgres/internals.go
+```
+
+You should still be using wit-bindgen v0.2.0 to generate the C bindings, this method is only useful for generating various Go types at the moment and does still require some manual editing because of a few bugs in the tiny-go generator and ABI incompatibilities between WASI preview1 and preview2.
+
+### Fix generated output
+
+It does take a little bit of effort to fix names to match intended output. Here are some commands that can assist in the renaming process:
+
+```shell
+# change the go package name to match your folder
+sed -I '' 's/package outbound_pg/package postgres/g' ./sdk/go/postgres/internals.go
+
+# fix a difference in header naming between v0.2.0
+sed -I '' 's/outbound_pg.h/outbound-pg.h/g' ./sdk/go/postgres/internals.go
+
+# remove everything for the generated world, there are bugs in the implementation so this part is easier to just write from scratch
+# this only works because it's generated last in the file, use your own discretion here
+sed -I '' '/type OutboundPg/,$d' ./sdk/go/postgres/internals.go
+
+# the function is the only place using unsafe code so we remove the import
+sed -I '' 's/import "unsafe"//' ./sdk/go/postgres/internals.go
+
+# remove the prefixes generated by importing the interfaces in wit world
+sed -I '' 's/PgTypes//g' ./sdk/go/postgres/internals.go
+sed -I '' 's/RdbmsTypes//g' ./sdk/go/postgres/internals.go
+```
+
+At this point, you should be seeing an error for the undefined functions removed and can start implementing those. This is at-best a hack so you don't have to write *as much* CGo marshalling code but you do still need to write the functions because there are some bugs in the generator.
